### 2018-10-10 #1010 @whedon
[PRE REVIEW]: TipToft: detecting plasmids contained in uncorrected long read sequencing data
**Submitting author:** @andrewjpage (<a href="http://orcid.org/0000-0001-6919-6062">Andrew Page</a>)
**Repository:** <a href="https://github.com/andrewjpage/tiptoft" target ="_blank">https://github.com/andrewjpage/tiptoft</a>
**Version:** v0.1.4
