### 2018-10-10 #1010 @arfon
@mgymrek - would you be willing to edit this submission for JOSS?
