+++ MANIFEST.in
@@ -1,1 +1,3 @@
 recursive-include cython *
+include VERSION
+include LICENSE
