### 2018-10-11 #1010 @kapsakcj
I'd also be happy to review. Thanks for nominating me to review @andrewjpage !  I will say though, that this would be my first time formally reviewing a manuscript and first time reviewing open-source software, so any advice or instructions would be welcomed. I have found the JOSS reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) to get me started. If you'd like to choose a more experienced reviewer, that would be fine with me.
