### 2018-10-12 #1010 @danielskatz
For me, the number depends on my expertise on the topic
