### 2018-10-15 #1010 @azneto
I'd be happy to do it, though it's my first time reviewing a submission for JOSS. Feel free to choose a more experienced reviewer.
