### 2018-10-15 #1010 @Kevin-Mattheus-Moerman
Thanks @azneto that is fine. prior experience with JOSS is not a requirement. I'm here to guide you through the process.
