### 2018-10-15 #1021 @whedon
[REVIEW]: TipToft: detecting plasmids contained in uncorrected long read sequencing data
**Submitting author:** @andrewjpage (<a href="http://orcid.org/0000-0001-6919-6062">Andrew Page</a>)
**Repository:** <a href="https://github.com/andrewjpage/tiptoft" target ="_blank">https://github.com/andrewjpage/tiptoft</a>
**Version:** v1.0.1
