### 2018-11-09 #1021 @kapsakcj
Hi everyone, I apologize for the delay in reviewing, but I've finally found a chance to give TipToft a proper review. I'm very impressed with TipToft and will likely eventually incorporate it into the software pipelines at where I work (public health lab). Thank you for developing it! I'm your target audience! We'll be doing nanopore sequencing soon and TipToft will be perfect.

It's not so often that you find a piece of bioinformatics software that is well-documented, easy to install(!), easy to use, and runs fast like the way TipToft does. I've installed/ran TipToft via pip, bioconda, docker, and even bioconda via Ubuntu via Windows Subsystem for Linux on Win10. Each worked on the first time too 👍

I tested TipToft on the supplied test data (got the expected output), some of my own nanopore data, and some random nanopore data I pulled off of the SRA. It ID'd all replicons that I expected except for one. I'd previously used PlasmidFinder to ID an IncA/C2 replicon off of a nanopore-read-only assembly of a plasmid, but was unable to do so using those same raw reads and TipToft. I'm thinking this may because these aren't the highest quality nanopore reads and isn't the fault of TipToft. @andrewjpage if you'd like I can probably supply you with the reads and assembly (need to check with PI) if you're curious to figure out what happened.

I do have one recommendation for the paper. I see the citation of the PlasmidFinder database in the repo, but I think it would be worth mentioning and citing the paper there too.

I think that's all I can suggest for right now. Thanks for developing the tool and publishing it in an open-access way. It was fun to review! I'll leave an issue on github if I have any, and potentially a PR if there's any way I can improve it. That's an ACCEPT from me.
