### 2018-11-09 #1021 @azneto
@andrewjpage, congratulations! This is a very very clever method for finding the sequences of interests in datasets with considerable error rates. TipToft relies on the kmer matching approach to identify perfect hits, without the hassle of dealing with base quality and sequence assembly. I've ran the software a few times during the last weeks using both Fedora and Ubuntu. The software is well documented and runs really fast.

Here are a few recommendations:

Though the documentation for the end users is really good and comprehensive, the code itself is not so inviting. I tried to follow the algorithm step by step, but it was really hard since it doesn't follow the PEP8 (Style Guide for Python Code) and the PEP20 (The Zen of Python).

I don't understand why would someone need to create aliases to functions or have duplicate functions in the code. For example:

Fasta.sequence_to_kmers is an alias to Fasta.sequence_kmers
Fasta.kmers_to_genes is an alias to Fasta.all_kmers_to_sq_in_file
The functions Fasta.sequence_kmers_vals and Fasta.sequence_kmers are exactly the same

There are variables and libraries that were created/loaded and never used (eg. time, operator, retcode). The software flake8 can find those and also check if the code is compatible with the PEP8. I also recommend using typing (https://docs.python.org/3/library/typing.html) for your functions to make sure the parameters and returns are correct and avoid implicitly passing of parameters using global variables.

Maybe it's worth mentioning in the paper that if the user doesn't have at least 20x sequencing coverage, it's almost certain TipToft will not find all kmers. In anyways, the sequencing companies recommend at least 40x coverage to guarantee full error correction and rarely someone will have such problem.

I agree with @hapsakcj regarding to the citation of PlasmidFinder and created an issue in the TipToft repo.

Thanks for contributing to the open source software and bioinformatics communities! I'm glad I got the chance to learn about a new useful tool. That's an ACCEPT for me too.
