+++ tiptoft/tests/Gene_test.py
@@ -10,22 +10,22 @@ class TestGene(unittest.TestCase):
         self.assertTrue(g.is_full_coverage())
         self.assertEqual(str(
             g),
-            'rep5.1    Full    100    NC005011    plasmidfinder'
-            '    rep5.1_rep(pMW2)_NC005011')
+            'rep5.1	Full	100	NC005011	plasmidfinder'
+            '	rep5.1_rep(pMW2)_NC005011')	
 
     def test_no_coverage(self):
         g = Gene('rep5.1_rep(pMW2)_NC005011', 0, 10)
         self.assertFalse(g.is_full_coverage())
         self.assertEqual(str(
-            g), 'rep5.1    Partial    0    NC005011    plasmidfinder'
-            '    rep5.1_rep(pMW2)_NC005011')
+            g), 'rep5.1	Partial	0	NC005011	plasmidfinder'
+            '	rep5.1_rep(pMW2)_NC005011')
 
     def test_medium_coverage(self):
         g = Gene('rep5.1_rep(pMW2)_NC005011', 5, 5)
         self.assertFalse(g.is_full_coverage())
         self.assertEqual(str(
-            g), 'rep5.1    Partial    50    NC005011    plasmidfinder'
-            '    rep5.1_rep(pMW2)_NC005011')
+            g), 'rep5.1	Partial	50	NC005011	plasmidfinder'
+            '	rep5.1_rep(pMW2)_NC005011')
 
     def test_nc_name(self):
         g = Gene('Col(BS512).1__NC_010656.2', 5, 5)
