### 2018-11-14 7aa8b3c
add plasmidfinder reference
