+++ paper.bib
@@ -63,3 +63,16 @@
   title = {Canu: scalable and accurate long-read assembly via adaptive k-mer weighting and repeat separation},
   journal = {Genome Research}
 }
+@article{Carattoli2014,
+  doi = {10.1128/aac.02412-14},
+  url = {https://doi.org/10.1128/aac.02412-14},
+  year  = {2014},
+  month = {apr},
+  publisher = {American Society for Microbiology},
+  volume = {58},
+  number = {7},
+  pages = {3895--3903},
+  author = {Alessandra Carattoli and Ea Zankari and Aurora Garc{\'{\i}}a-Fern{\'{a}}ndez and Mette Voldby Larsen and Ole Lund and Laura Villa and Frank M{\o}ller Aarestrup and Henrik Hasman},
+  title = {In {SilicoDetection} and Typing of Plasmids using {PlasmidFinder} and Plasmid Multilocus Sequence Typing},
+  journal = {Antimicrobial Agents and Chemotherapy}
+}
