+++ .travis.yml
@@ -1,5 +1,8 @@
 language: python
-python:
+python:\
+  - "3.2"
+  - "3.3"
+  - "3.4"
   - "3.5"
   - "3.6"
 sudo: false
