+++ .travis.yml
@@ -1,6 +1,5 @@
 language: python
-python:\
-  - "3.2"
+python:
   - "3.3"
   - "3.4"
   - "3.5"
