### 2018-11-14 2875a8b
remove redundant code
