+++ tiptoft/tests/Kmers_test.py
@@ -1,6 +1,5 @@
 import unittest
 import os
-import logging
 from tiptoft.Kmers import Kmers
 
 test_modules_dir = os.path.dirname(os.path.realpath(__file__))
