+++ tiptoft/tests/Blocks_test.py
@@ -1,5 +1,4 @@
 import unittest
-import os
 from tiptoft.Blocks import Blocks
 
 
