+++ tiptoft/Fasta.py
@@ -17,23 +17,22 @@ class Fasta:
         self.homopolyer_compression = homopolyer_compression
         self.max_kmer_count = max_kmer_count
 
-        self.sequences_to_kmers = self.sequence_kmers()
+        self.sequences_to_kmers = self.sequence_kmers('get_all_kmers_counter')
+        self.sequences_to_kmers_count =\
+            self.sequence_kmers('get_all_kmers_freq')
         self.all_kmers = self.all_kmers_in_file()
         self.kmers_to_genes = self.all_kmers_to_seq_in_file()
         self.kmer_keys_set = set(self.all_kmers.keys())
-        self.sequences_to_kmers_count = self.sequence_kmers()
 
     '''Count the kmers in a sequence'''
 
-    def sequence_kmers(self):
+    def sequence_kmers(self, kmer_action='get_all_kmers_counter'):
         seq_counter = 0
-
         kmer_to_sequences = {}
         for record in SeqIO.parse(self.filename, "fasta"):
-
             kmers = Kmers(str(record.seq), self.k, self.homopolyer_compression)
             # We assume here that the sequence name is unique in the FASTA file
-            kmer_to_sequences[record.id] = kmers.get_all_kmers_counter(
+            kmer_to_sequences[record.id] = getattr(kmers, kmer_action)(
                 max_kmer_count=self.max_kmer_count)
 
             seq_counter += 1
