### 2018-11-14 #1021 @andrewjpage
Thank you @kapsakcj and @azneto for the kind reviews and for taking the time to use and review the code.

I have added in the PlasmidFinder reference and mentioned it in the text.

I have reformatted the code and removed anything thats redundant or duplicated, so it now passes flake8 and pycodestyle without any errors. Thank you for the pointer to PEP20, its the first time I've come across it and I most definitely will follow it for all new Python projects. I'm relatively new to Python and have unfortunatly carried over bad habits from other languages.

```
I don't understand why would someone need to create aliases to functions. For example:
Fasta.sequence_to_kmers is an alias to Fasta.sequence_kmers
Fasta.kmers_to_genes is an alias to Fasta.all_kmers_to_sq_in_file
```
In this instance I wanted to run a computationally intensive method exactly once and store the result. Happy to discuss this offline if there is a more appropriate pattern to use in Python.

```
I also recommend using typing (https://docs.python.org/3/library/typing.html) for your functions to make sure the parameters and returns are correct and avoid implicitly passing of parameters using global variables.
```
Type hints look really awesome, thank for the suggestion.  As this software currently works on 3.4 and above, adding type hints would increase this to 3.5 (probably 3.6 given the provisional status of the api). I think I will wait for it to stabilise before making the switch to minimise any potential barriers for researchers.

```
Maybe it's worth mentioning in the paper that if the user doesn't have at least 20x sequencing coverage, it's almost certain TipToft will not find all kmers. In anyways, the sequencing companies recommend at least 40x coverage to guarantee full error correction and rarely someone will have such problem.
```

Yes you are absolutely right, in the real world you may not see all kmers if the coverage is too low. As its essentially an alignment based approach rather than de novo assembly it can detect a signal at a low coverage. I have added text to explain the depth of coverage. Unfortunatly there is not enough room in a JOSS paper to explain the algorithm indepth.
