### 2018-11-14 7012930
update paper and version
