### 2019-01-16 andrewjpage/tiptoft#17 @ctb
create a CITATION file and add plasmidFinder to it?
One way to help users cite the appropriate literate is to create a CITATION.cff file, https://citation-file-format.github.io/. (there are other approaches, too, of course ;). ref #16.

This is just a suggestion, not critical :)
