### 2019-01-16 andrewjpage/tiptoft#18 @ctb
cannot find example data file
the link [here](https://github.com/andrewjpage/tiptoft#example-usage) to an example data file does not work & the data file is not included in the repo any more.

I found the commit that deleted it like so,

```
git rev-list -n 1 HEAD -- ERS654932_plasmids.fastq.gz
```

and then checked it out like so,

```
git show 7c7d3e55da84cf814c783dbaf429def261c77328^:example_data/ERS654932_plasmids.fastq.gz > ERS654932_plasmids.fastq.gz
```

I think it'd be great if the file remained in the repo, but it should be provided somewhere (or removed from the README - -1 on that :). Note that deleting it without rewriting history means it's still in the git history so it doesn't speed up git clones, although it does reduce the bundled repo size for releases etc.
