### 2019-01-16 #1021 @ctb
Review done! Agree with all the nice comments above - good work!

I did not look at the code at all, since @azneto did such a nice job :)

The only two issues that I saw that should be addressed are indicated in my unchecked boxes above --

- the version has advanced to 1.0! so I left unchecked the "version matches the released version" box; minor technicality since the version has surpassed that :).
- the example data file mentioned in the README no longer exists in the repo. See issue https://github.com/andrewjpage/tiptoft/issues/18.

but I did file another two issues that are merely suggestions,

- maybe create a CITATION.cff file containing plasmidFinder? https://github.com/andrewjpage/tiptoft/issues/17
- it would be useful to have an explicit CONTRIBUTING.md https://github.com/andrewjpage/tiptoft/issues/18

Once the example data file issue is resolved (one way or another) I'm definitely an ACCEPT!
