### 2019-01-16 #1021 @Kevin-Mattheus-Moerman
@andrewjpage please reply to @ctb 's comments. Please do add a CONTRIBUTING.md file (see also https://help.github.com/articles/setting-guidelines-for-repository-contributors/).

@ctb feel free to tick that version box. We will update the final version number upon acceptance (future review issues will not have the version box any more as versions often move on).
