### 2019-01-18 andrewjpage/tiptoft#17 @andrewjpage
Added. Thanks for the suggestion, didnt know the format existed, so must roll it out to all my repos.
