### 2019-01-18 andrewjpage/tiptoft#18 @andrewjpage
Good catch!  I've added in a shrunken example file, which produces the same example output.
