### 2019-01-18 #1021 @andrewjpage
Thanks @ctb for taking the time to review the software. I have resolved all the issues raised, adding a CITATION.cff file, added back the example data file, adding a CONTRIBUTING.md  and a CoC.

Thanks @Kevin-Mattheus-Moerman for editing.
