### 2019-02-01 2615f76
Update paper.md
editorial typo fix
