### 2019-02-10 #1021 @Kevin-Mattheus-Moerman
Okay it looks like we are in good shape. @andrewjpage can you:

- [ ] create an archived version of the reviewed software on a service like ZENODO (some use these steps https://guides.github.com/activities/citable-code/) and quote the DOI here
- [ ] List the latest version/tag number here of the reviewed software
