### 2019-02-10 #1021 @andrewjpage
The Zenodo DOI for the reviewed software is here:
http://doi.org/10.5281/zenodo.2561192

The latest version of the reviewed software is:
v1.0.1
