### 2019-03-01 #1021 @danielskatz
I suspect an error in the whedon's accept processing.  I will check by generating the pdf by itself
