### 2019-03-01 #1021 @danielskatz
👋 @arfon - my issues seem to be resolved - should I go ahead and finalize the accept?
(Also, any idea what happened with the generate process as part of accept, and the faulty PDF?)
