### 2019-03-01 #1021 @arfon
> 👋 @arfon - my issues seem to be resolved - should I go ahead and finalize the accept

Yes, please do.

Sometimes Whedon fails to compile the PDF properly - this bug is something I've never managed to replicate.

As for the faulty PDF, `@whedon generate pdf` doesn't attempt to make the correct link for the archive (apologies, this is undocumented).
